+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+### v0.2.0
+* Implemented start menu
+* Implemented preview window and drop ability
+* Added music, sound effect, and font assets
+
 ### v0.1.1 
 * Eased difficulty scaling
 * Added outlines to line clears